auth-backend: move a few types to auth-node
